pim6d: Implementing "show ipv6 mroute" CLI
Adding new show CLI to display ipv6 mroute information.
Moved show_mroute function from pim_cmd.c to pim_cmd_common.c.
